requirements.txt
pandas
matplotlib
seaborn
python-dotenv
pydantic
duckdb>=0.9.0
ragas
streamlit
chromadb>=1.4.0
tiktoken
prompty
openai
langchain
langchain-core
langchain-community
langchain-openai
langchain-prompty
langchain-experimental>=0.0.1
langchain-chroma
promptflow
opentelemetry-api
opentelemetry-sdk
openinference-instrumentation-langchain
azure-ai-evaluation
azure-identity
azure-search-documents


# Dev Dependencies - Code Quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
flake8>=6.0.0
pytest>=7.0.0

